initial upload
Simple scripts to convert in-situ observations to (1) the format
required by ILAMB V1,  then to (2) ILAMB V2. Because I started to use
V1 firstly. The benchmark data is already at ILAMB V1. I just wrote a
script convert them from V1 to V2. The code is not smart. Need to be
improved much more. An example of SNOW DEPTH (snd) is attached here.